Added simple http server and removed unused classes
