--- src/ktt/lib/httpserver/server/HttpPrincipal.java
@@ -1,5 +1,0 @@
-package ktt.lib.httpserver.server;
-
-@Deprecated
-public abstract class HttpPrincipal {
-}
